[REFACTOR] 메인 화면 모델, 인터랙션 리팩토링
