+++ app/src/main/java/com/polzzak_android/presentation/main/intercation/MainProgressInteraction.kt
@@ -1,14 +1,14 @@
-package com.polzzak_android.presentation.main.protector.progress
+package com.polzzak_android.presentation.main.intercation
 
 import android.widget.TextView
 import androidx.viewpager2.widget.ViewPager2
-import com.polzzak_android.presentation.main.protector.model.StampBoardSummary
+import com.polzzak_android.presentation.main.model.StampBoardSummary
 import com.polzzak_android.presentation.component.SemiCircleProgressView
 
-interface ProgressInteraction {
+interface MainProgressInteraction {
     fun setViewPager(view: ViewPager2, curInd: TextView, totalInd: TextView, stampList: List<StampBoardSummary>)
 
     fun onStampPagerClicked(stampBoardItem: StampBoardSummary)
 
-    fun setProgressAnim(view: SemiCircleProgressView)
+    fun setProgressAnim(curCnt: Int, totalCnt: Int, view: SemiCircleProgressView)
 }
