+++ app/src/main/java/com/polzzak_android/presentation/main/intercation/MainSelectUserInteraction.kt
@@ -1,0 +1,6 @@
+package com.polzzak_android.presentation.main.intercation
+
+
+interface MainSelectUserInteraction {
+    fun onUserClicked(userItem: String)
+}
