+++ app/src/main/java/com/polzzak_android/presentation/main/model/StampBoardSummary.kt
@@ -1,4 +1,4 @@
-package com.polzzak_android.presentation.main.protector.model
+package com.polzzak_android.presentation.main.model
 
 data class StampBoardSummary(
     val currentStampCount: Int,
