--- app/src/main/java/com/polzzak_android/presentation/main/protector/progress/SelectUserInteraction.kt
@@ -1,6 +1,0 @@
-package com.polzzak_android.presentation.main.protector.progress
-
-
-interface SelectUserInteraction {
-    fun onUserClicked(userItem: String)
-}
